Asynchronous creation of mesh tile
and few fixes
